[MP] Some 64 bit fixes for module API

+++ codemp/cgame/cg_main.c
@@ -3162,7 +3162,7 @@ cgameImport_t *trap = NULL;
 Q_EXPORT cgameExport_t* QDECL GetModuleAPI( int apiVersion, cgameImport_t *import )
 {
 	static cgameExport_t cge = {0};
-	
+
 	assert( import );
 	trap = import;
 	Com_Printf	= trap->Print;
@@ -3203,7 +3203,6 @@ Q_EXPORT cgameExport_t* QDECL GetModuleAPI( int apiVersion, cgameImport_t *impor
 	cge.MiscEnt					= CG_MiscEnt;
 	cge.CameraShake				= CG_FX_CameraShake;
 
-
 	return &cge;
 }
 
@@ -3215,10 +3214,10 @@ This is the only way control passes into the module.
 This must be the very first function compiled into the .q3vm file
 ================
 */
-Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
-
+Q_EXPORT intptr_t vmMain( int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4,
+	intptr_t arg5, intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11 )
+{
 	switch ( command ) {
-
 	case CG_INIT:
 		CG_Init( arg0, arg1, arg2 );
 		return 0;
@@ -3264,12 +3263,12 @@ Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, i
 		return 0;
 
 	case CG_GET_GHOUL2:
-		return (int)cg_entities[arg0].ghoul2; //NOTE: This is used by the effect bolting which is actually not used at all.
+		return (intptr_t)cg_entities[arg0].ghoul2; //NOTE: This is used by the effect bolting which is actually not used at all.
 											  //I'm fairly sure if you try to use it with vm's it will just give you total
 											  //garbage. In other words, use at your own risk.
 
 	case CG_GET_MODEL_LIST:
-		return (int)cgs.gameModels;
+		return (intptr_t)cgs.gameModels;
 
 	case CG_CALC_LERP_POSITIONS:
 		CG_CalcEntityLerpPositions( &cg_entities[arg0] );
@@ -3278,8 +3277,10 @@ Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, i
 	case CG_TRACE:
 		C_Trace();
 		return 0;
+
 	case CG_GET_SORTED_FORCE_POWER:
 		return forcePowerSorted[arg0];
+
 	case CG_G2TRACE:
 		C_G2Trace();
 		return 0;
@@ -3306,10 +3307,10 @@ Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, i
 		return 0;
 
 	case CG_GET_ORIGIN_TRAJECTORY:
-		return (int)CG_GetOriginTrajectory( arg0 );
+		return (intptr_t)CG_GetOriginTrajectory( arg0 );
 
 	case CG_GET_ANGLE_TRAJECTORY:
-		return (int)CG_GetAngleTrajectory( arg0 );
+		return (intptr_t)CG_GetAngleTrajectory( arg0 );
 
 	case CG_ROFF_NOTETRACK_CALLBACK:
 		_CG_ROFF_NotetrackCallback( arg0, (const char *)arg1 );
