+++ codemp/ui/ui_main.c
@@ -10507,13 +10507,15 @@ vmMain
 ============
 */
 
-Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
+Q_EXPORT intptr_t vmMain( int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4,
+	intptr_t arg5, intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11 )
+{
 	switch ( command ) {
 	case UI_GETAPIVERSION:
 		return UI_LEGACY_API_VERSION;
 
 	case UI_INIT:
-		UI_Init(arg0);
+		UI_Init( arg0 );
 		return 0;
 
 	case UI_SHUTDOWN:
